Core run loop
Is able to take a program and transform it. At the end of the program,
it prints the stack.

Notably, the echo command is not yet implemented. It was unclear how to
cause an operation to print something to the console without "tainting"
the main loop with the IO monad.